Add separate file for command line options and arguments.
The great argparse module is used here.
It allows to easily parse various combinations of options and arguments from the command line.
First of all, user can specify a port number to start a listener on.
Secondly, the user can turn the verbosity on and off.
Also, argparse generates nice help and usage messages for free.