Rename runme.py to mfhclient.py
As we`re going to have a client-server architecture, we split the program
into 2 pieces.
mfhclient.py is a client and server.py is a server.
The idea behind this is to be able to connect many clients to the server
and make the communicate nicely.
Also, we`re going to have a launcher, so that instead of starting client
and server separately, you just specify an option to the launcher.