Add launcher for client with ability to self-update.
The launcher starts 1 thread with client and 1 thread with updater.
For now, updates are made in time intervals of 1 hour.
Updates are just pulls from git repository.

When the time comes, main thread passes the "quit" message to client
thread and it shuts itself down.
Although, the client thread shuts itself down not instantly, but after a
request has been made to it.
This is planned to be fixed by moving to multiprocessing, where you
actually can send a SIGINT of SIGTERM to the process and it will shut
itself down.
You can`t do this with threads, though.

Then, when the thread is dead, main thread launches updater thread, which
does a git pull and pip install from requirements.txt.
With these actions performed, we will be able to get the latest code and
install all the new requirements, which may have been added.

Then the whole thing shuts down and starts itself as a new process,
overwriting the old process with arguments, which were specified to the
old process. And the new code runs gracefully.
Pretty cool, huh?