Switch from multithreading to multiprocessing
It is easier to work with processes in terms of the current task.
The task was to make the client and updater work at the same time.
Threads have a lot of limitations and are harder to get right, so I
switched to processes.
The code went pretty much straightforward.
Not on the first try, though.