Add argument for launching client
Before, client would launch automatically on port you specify as a
positional argument.
This means you can't just not to launch the client.
Now, there are 2 options:
    1) -c to launch client on the port defined in settings
    2) --client to specify the port explicitly
So, if you want to launch just the server without client, there
will be an option for that, just say what to launch.