Make a proper file structure
It was decided to make a file structure for the project, because
there was just so many files in the root folder it got messy.
Now they are placed in a somewhat logical order. This is not final
and may change.
Also, with such a file structure it is easier to make a package to
be able to install it from pip(which is a milestone).