Fix a bug with launching mfh from another directory
There are some places when we use relative paths in the project.
In this case it is copying setting example to setting.py.
If you lauched it from a directory, which is not a directory where
mfh.py resides, you would get IOError, because python prepended a
relative path to your current directory, where it is very unlikely
to find settings files.